Remove `tz.TimeZone.Offset`; split up `offsetAtTimestamp`
I was planning on doing this, but then decided that not to when I
realized that it would introduce redundant computations. However! While
I was implementing the Win32 TimeZone, I realized getting the
designation was fairly involved, so I decided to go ahead with splitting
up the `offsetAtTimestamp` function into `offsetAtTimestamp`,
`isDaylightSavingTimeAtTimestamp`, and `designationAtTimestamp`.